pyproject.toml
[tool.poetry]
name = "obsync"
version = "0.1.0"
description = ""
authors = ["ShiinaRinne"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = "^0.24.0.post1"
jwt = "^1.3.1"
pymysql = "^1.1.0"
bcrypt = "^4.0.1"
sqlalchemy = "^2.0.23"
python-jose = "^3.3.0"
loguru = "^0.7.2"
websockets = "^12.0"
requests = "^2.31.0"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
